Testing agiel

+++ Server/services/app/models/class.js
@@ -61,7 +61,7 @@ module.exports = (sequelize, DataTypes) => {
           },
         },
       },
-      averageRating: DataTypes.INTEGER,
+      averageRating: DataTypes.DECIMAL,
       status: DataTypes.STRING,
       SubjectId: DataTypes.INTEGER,
       description: {
